requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
dj-config-url
psycopg2-binary
twilio
celery[redis]
drf-yasg
django-cors-headers
djangochannelsrestframework